requirements.txt
torch
pyyaml
matplotlib
numpy
Pillow
tqdm